skyforge/apps/api/pyproject.toml
[project]
name = "skyforge-api"
version = "0.1.0"
description = "Skyforge API — drone data processing backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.15.0",
    "asyncpg>=0.30.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-pubsub>=2.28.0",
    "google-auth>=2.38.0",
    "firebase-admin>=6.6.0",
    "google-generativeai>=0.8.0",
    "stripe>=12.0.0",
    "alembic>=1.14.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.0",
    "python-multipart>=0.0.20",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
